refactoring(15): extract 5 per-tab subcomponents from DetailPanel
Move the trace/code/memory/access/report tab render blocks out of
DetailPanel.tsx into DetailPanelTabs/{Trace,Code,Memory,Access,Report}Tab.tsx
with explicit prop contracts (t sourced via useI18n internally; node.id
passed as currentNodeId). JSX moved verbatim; the activeTab=== guards stay
in the container. DetailPanel 794->425 lines (938->425 over req 15).
Behavior-equivalence adversarially verified per tab + container wiring
(6 independent skeptics, all equivalent, zero discrepancies).
