+++ ari-core/ari/viz/frontend/src/components/README.md
@@ -73,7 +73,12 @@ Feature-grouped React components — one subdirectory per dashboard page plus sh
   - `useDetailPanelData.ts` — hook owning DetailPanel's memory/access-log/node-report fetch effects; extracted from DetailPanel in req 15.
   - `DetailPanelTabs/` — extracted detail-panel subcomponents.
     - `README.md` — DetailPanelTabs index.
+    - `AccessTab.tsx` — memory access-log tab (write/read events); extracted from DetailPanel in req 15.
+    - `CodeTab.tsx` — code-snippets tab (run_code from trace_log); extracted from DetailPanel in req 15.
     - `MemoryEntryCard.tsx` — renders one memory record (own/inherited/global) as a card.
+    - `MemoryTab.tsx` — memory tab (own/inherited/global entry cards); extracted from DetailPanel in req 15.
+    - `ReportTab.tsx` — node-report tab (node_report.json structured view); extracted from DetailPanel in req 15.
+    - `TraceTab.tsx` — MCP-trace tab (tool pills + colored trace log); extracted from DetailPanel in req 15.
 - `Wizard/` — multi-step run-launch wizard.
   - `README.md` — Wizard index.
   - `index.ts` — barrel re-export.
