+++ ari-core/ari/viz/frontend/src/components/Tree/DetailPanelTabs/README.md
@@ -5,4 +5,9 @@ Leaf subcomponents extracted from `DetailPanel.tsx` for the node-detail tabs.
 ## Contents
 
 - `README.md` — this file.
+- `AccessTab.tsx` — memory access-log tab (write/read events); extracted from DetailPanel in req 15.
+- `CodeTab.tsx` — code-snippets tab (run_code from trace_log); extracted from DetailPanel in req 15.
 - `MemoryEntryCard.tsx` — renders one memory record (own/inherited/global) as a card.
+- `MemoryTab.tsx` — memory tab (own/inherited/global entry cards); extracted from DetailPanel in req 15.
+- `ReportTab.tsx` — node-report tab (node_report.json structured view); extracted from DetailPanel in req 15.
+- `TraceTab.tsx` — MCP-trace tab (tool pills + colored trace log); extracted from DetailPanel in req 15.
